Created Layout namespace with functions for setting up Window panels.
Removed main.h file
Reorganized include dependencies
