+++ GLUT_Window_Manager/glut_main.cpp
@@ -1,29 +1,24 @@
+#include "Window.h"
+#include "Layout.h"
+#include "Panel.h"
+#include "GraphicsManager.h"
+#include "Command.h"
 #include <GLEW/glew.h>
 #include <GLUT/freeglut.h>
-
-#include <stdio.h>
-#include <iostream>
-#include <ostream>
-#include <fstream>
-#include <time.h>
-#include <algorithm>
-
-#include "main.h"
-#include "Window.h"
-#include "Command.h"
 
 int main(int argc, char **argv)
 {
     Panel* windowPanel = Window::Instance().GetWindowPanel();
 
-    SetUpWindow(*windowPanel);
+    Layout::SetUpWindow(*windowPanel);
+    GraphicsManager* graphicsManager = windowPanel->GetPanel("Graphics")->GetGraphicsManager();
 
     Window::Instance().InitializeGLUT(argc, argv);
     Window::Instance().InitializeGL();
 
-    SetUpGLInterop(*windowPanel);
-    SetUpCUDA(*windowPanel);
-    SetUpShaders(*windowPanel);
+    graphicsManager->SetUpGLInterop();
+    graphicsManager->SetUpCuda();
+    graphicsManager->SetUpShaders();
 
     Window::Instance().Display();
