+++ LBM_GL_CPP/Panel.cpp
@@ -1,8 +1,8 @@
 #include "Panel.h"
 #include "GraphicsManager.h"
 #include "Mouse.h"
-#include <algorithm>
 #include <GLEW/glew.h>
+#include <algorithm>
 
 Color::Color()
 {
